+++ src/apis/Profile/getProfile.js
@@ -1,4 +1,4 @@
-import axiosInstance from './axiosInstance';
+import axiosInstance from '../axiosInstance';
 
 export default async function getProfile(userId) {
   const { data: { data } } = await axiosInstance.get(`/profiles/${userId}`);
