+++ src/apis/index.js
@@ -1,6 +1,6 @@
 export { default as api } from './axiosInstance';
-export { default as askQuestion } from './askQuestion';
-export { default as getTaskList } from './getTaskList';
+export { default as askQuestion } from './Task/askQuestion';
+export { default as getTaskList } from './Task/getTaskList';
 export { default as useAuth } from './useAuth';
-export { default as getProfile } from './getProfile';
+export { default as getProfile } from './Profile/getProfile';
